refactor(main): extract command handlers into cmd/ modules (#90)
